Add support for milliseconds from epoch
The Handlebars helper for npmap.js could not accept milliseconds
from epoch, as the string was passed as a date string directly
to the Date constructor.

To allow for time from epoch, added an additional option
to the date formatter to specify the time as epoch or as a
date string. If epoch, then the string is converted to an
int and passed to the Date constructor.

Updated the test.js function to match this, and added a test to
verify that the epoch time was being properly parsed.